Agrego test de ubicacion y modifico los de muestra

+++ src/src/Ubicacion.java
@@ -1,44 +1,34 @@
 package src;
 
-
-
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class Ubicacion {
 
-	private Integer latitud;
-	private Integer longitud;
+	private Double latitud;
+	private Double longitud;
 	
-	public Ubicacion(Integer latitud, Integer longitud){
+	public Ubicacion(Double latitud, Double longitud){
 		this.latitud=latitud;
 		this.longitud=longitud;
 	}
 	
 	
-	
-	public Integer getLatitud() {
+	public Double getLatitud() {
 		return latitud;
 	}
 
-
-
-	public void setLatitud(Integer latitud) {
+	public void setLatitud(Double latitud) {
 		this.latitud = latitud;
 	}
 
-
-
-	public Integer getLongitud() {
+	public Double getLongitud() {
 		return longitud;
 	}
 
-
-
-	public void setLongitud(Integer longitud) {
+	public void setLongitud(Double longitud) {
 		this.longitud = longitud;
 	}
-
 
 
 	/*
